chore: refactor to tower layer
