+++ crates/common/src/provider/mod.rs
@@ -4,3 +4,4 @@ pub mod alloy;
 pub mod ethers;
 pub mod retry;
 pub mod runtime_transport;
+pub mod tower;
