+++ crates/common/src/provider/runtime_transport.rs
@@ -1,7 +1,6 @@
 //! Runtime transport that connects on first request, which can take either of an HTTP,
 //! WebSocket, or IPC transport and supports retries based on CUPS logic.
-use super::retry::{RateLimitRetryPolicy, RetryPolicy};
-use crate::{ALCHEMY_FREE_TIER_CUPS, REQUEST_TIMEOUT};
+use crate::REQUEST_TIMEOUT;
 use alloy_json_rpc::{RequestPacket, ResponsePacket};
 use alloy_pubsub::{PubSubConnect, PubSubFrontend};
 use alloy_transport::{
@@ -12,14 +11,7 @@ use alloy_transport_ipc::IpcConnect;
 use alloy_transport_ws::WsConnect;
 use ethers_providers::{JwtAuth, JwtKey};
 use reqwest::header::{HeaderName, HeaderValue};
-use std::{
-    path::PathBuf,
-    str::FromStr,
-    sync::{
-        atomic::{AtomicU32, Ordering},
-        Arc,
-    },
-};
+use std::{path::PathBuf, str::FromStr, sync::Arc};
 use thiserror::Error;
 use tokio::sync::RwLock;
 use tower::Service;
@@ -86,18 +78,6 @@ pub struct RuntimeTransport {
     jwt: Option<String>,
     /// The timeout for requests.
     timeout: std::time::Duration,
-    /// The current amount of requests being retried.
-    requests_enqueued: Arc<AtomicU32>,
-    /// The policy that will be used to determine whether to retry a request or not.
-    policy: RateLimitRetryPolicy,
-    /// Max amount of retries.
-    max_rate_limit_retries: u32,
-    /// Max amount of timeout retries
-    max_timeout_retries: u32,
-    /// Initial backoff
-    initial_backoff: u64,
-    /// Compute units per second
-    compute_units_per_second: u64,
 }
 
 /// A builder for [RuntimeTransport].
@@ -107,25 +87,12 @@ pub struct RuntimeTransportBuilder {
     headers: Vec<String>,
     jwt: Option<String>,
     timeout: std::time::Duration,
-    max_rate_limit_retries: u32,
-    max_timeout_retries: u32,
-    initial_backoff: u64,
-    compute_units_per_second: u64,
 }
 
 impl RuntimeTransportBuilder {
     /// Create a new builder with the given URL.
     pub fn new(url: Url) -> Self {
-        Self {
-            url,
-            headers: vec![],
-            jwt: None,
-            timeout: REQUEST_TIMEOUT,
-            max_rate_limit_retries: 8,
-            max_timeout_retries: 8,
-            initial_backoff: 800,
-            compute_units_per_second: ALCHEMY_FREE_TIER_CUPS,
-        }
+        Self { url, headers: vec![], jwt: None, timeout: REQUEST_TIMEOUT }
     }
 
     /// Set the URL for the transport.
@@ -146,30 +113,6 @@ impl RuntimeTransportBuilder {
         self
     }
 
-    /// Set the max rate-limit retries for the transport.
-    pub fn with_max_rate_limit_retries(mut self, max_retries: u32) -> Self {
-        self.max_rate_limit_retries = max_retries;
-        self
-    }
-
-    /// Set the max timeout retries for the transport.
-    pub fn with_max_timeout_retries(mut self, max_retries: u32) -> Self {
-        self.max_timeout_retries = max_retries;
-        self
-    }
-
-    /// Set the initial backoff for the transport.
-    pub fn with_initial_backoff(mut self, initial_backoff: u64) -> Self {
-        self.initial_backoff = initial_backoff;
-        self
-    }
-
-    /// Set the compute units per second for the transport.
-    pub fn with_cups(mut self, cups: u64) -> Self {
-        self.compute_units_per_second = cups;
-        self
-    }
-
     /// Builds the [RuntimeTransport] and returns it in a disconnected state.
     /// The runtime transport will then connect when the first request happens.
     pub fn build(self) -> RuntimeTransport {
@@ -179,12 +122,6 @@ impl RuntimeTransportBuilder {
             headers: self.headers,
             jwt: self.jwt,
             timeout: self.timeout,
-            requests_enqueued: Arc::new(AtomicU32::new(0)),
-            policy: RateLimitRetryPolicy,
-            max_rate_limit_retries: self.max_rate_limit_retries,
-            max_timeout_retries: self.max_timeout_retries,
-            initial_backoff: self.initial_backoff,
-            compute_units_per_second: self.compute_units_per_second,
         }
     }
 }
@@ -283,76 +220,12 @@ impl RuntimeTransport {
             // SAFETY: We just checked that the inner transport exists.
             let inner_mut = inner.as_mut().expect("We should have an inner transport.");
 
-            let ahead_in_queue = this.requests_enqueued.fetch_add(1, Ordering::SeqCst) as u64;
-
-            let mut rate_limit_retry_number: u32 = 0;
-            let mut timeout_retries: u32 = 0;
-
-            loop {
-                let err;
-                let fut = match inner_mut {
-                    InnerTransport::Http(http) => http.call(req.clone()),
-                    InnerTransport::Ws(ws) => ws.call(req.clone()),
-                    InnerTransport::Ipc(ipc) => ipc.call(req.clone()),
-                }
-                .await;
-
-                match fut {
-                    Ok(res) => {
-                        this.requests_enqueued.fetch_sub(1, Ordering::SeqCst);
-                        return Ok(res)
-                    }
-                    Err(e) => err = e,
-                }
-
-                let err = TransportError::from(err);
-                let should_retry = this.policy.should_retry(&err);
-                if should_retry {
-                    rate_limit_retry_number += 1;
-                    if rate_limit_retry_number > this.max_rate_limit_retries {
-                        return Err(TransportErrorKind::custom_str("Max retries exceeded"))
-                    }
-
-                    let current_queued_reqs = this.requests_enqueued.load(Ordering::SeqCst) as u64;
-
-                    // try to extract the requested backoff from the error or compute the next
-                    // backoff based on retry count
-                    let mut next_backoff = this
-                        .policy
-                        .backoff_hint(&err)
-                        .unwrap_or_else(|| std::time::Duration::from_millis(this.initial_backoff));
-
-                    // requests are usually weighted and can vary from 10 CU to several 100 CU,
-                    // cheaper requests are more common some example alchemy
-                    // weights:
-                    // - `eth_getStorageAt`: 17
-                    // - `eth_getBlockByNumber`: 16
-                    // - `eth_newFilter`: 20
-                    //
-                    // (coming from forking mode) assuming here that storage request will be the
-                    // driver for Rate limits we choose `17` as the average cost
-                    // of any request
-                    const AVG_COST: u64 = 17u64;
-                    let seconds_to_wait_for_compute_budget = compute_unit_offset_in_secs(
-                        AVG_COST,
-                        this.compute_units_per_second,
-                        current_queued_reqs,
-                        ahead_in_queue,
-                    );
-                    next_backoff +=
-                        std::time::Duration::from_secs(seconds_to_wait_for_compute_budget);
-
-                    tokio::time::sleep(next_backoff).await;
-                } else {
-                    if timeout_retries < this.max_timeout_retries {
-                        timeout_retries += 1;
-                        continue;
-                    }
-
-                    this.requests_enqueued.fetch_sub(1, Ordering::SeqCst);
-                    return Err(TransportErrorKind::custom_str("Max retries exceeded"))
-                }
+            match inner_mut {
+                InnerTransport::Http(http) => http.call(req.clone()),
+                InnerTransport::Ws(ws) => ws.call(req.clone()),
+                InnerTransport::Ipc(ipc) => ipc.call(req.clone()),
             }
+            .await
         })
     }
 
@@ -434,29 +307,4 @@ fn url_to_file_path(url: &Url) -> Result<PathBuf, ()> {
 #[cfg(not(windows))]
 fn url_to_file_path(url: &Url) -> Result<PathBuf, ()> {
     url.to_file_path()
-}
-
-/// Calculates an offset in seconds by taking into account the number of currently queued requests,
-/// number of requests that were ahead in the queue when the request was first issued, the average
-/// cost a weighted request (heuristic), and the number of available compute units per seconds.
-///
-/// Returns the number of seconds (the unit the remote endpoint measures compute budget) a request
-/// is supposed to wait to not get rate limited. The budget per second is
-/// `compute_units_per_second`, assuming an average cost of `avg_cost` this allows (in theory)
-/// `compute_units_per_second / avg_cost` requests per seconds without getting rate limited.
-/// By taking into account the number of concurrent request and the position in queue when the
-/// request was first issued and determine the number of seconds a request is supposed to wait, if
-/// at all
-fn compute_unit_offset_in_secs(
-    avg_cost: u64,
-    compute_units_per_second: u64,
-    current_queued_requests: u64,
-    ahead_in_queue: u64,
-) -> u64 {
-    let request_capacity_per_second = compute_units_per_second.saturating_div(avg_cost);
-    if current_queued_requests > request_capacity_per_second {
-        current_queued_requests.min(ahead_in_queue).saturating_div(request_capacity_per_second)
-    } else {
-        0
-    }
 }
